Add files via upload
Once again, messy (lots of test code). But the chart code is all there. I should have styled a lot of these more. Basically, once I had the underlying data structure set I just put up the x and y and the chart/graph. If it was obvious how to put a title and label the axes I did.